requirements.txt
alibabacloud_bssopenapi20171214==1.0.8
alibabacloud_tea_openapi==0.2.7
aliyun_python_sdk_core==2.13.36
elasticsearch==7.15.2
python_dateutil==2.8.2
requests==2.22.0
SoftLayer==6.1.1
tencentcloud_sdk_python_intl_en==3.0.332
aliyun-python-sdk-bssopenapi
